💥 Move some traits & actions from SDK to contracts
