+++ crates/contract-sdk/Cargo.toml
@@ -20,9 +20,6 @@ tracing = { version = "0.1", optional = true }
 risc0-zkvm = { version = "1.2.4", default-features = false, optional = true }
 sp1-zkvm = { version = "4.1.3", optional = true }
 
-[dev-dependencies]
-mockall = "0.13.0"
-
 [features]
 risc0 = ["dep:risc0-zkvm"]
 sp1 = ["dep:sp1-zkvm"]
