+++ crates/contracts/amm/src/lib.rs
@@ -2,7 +2,7 @@ use std::collections::BTreeMap;
 use std::hash::{Hash, Hasher};
 
 use borsh::{BorshDeserialize, BorshSerialize};
-use sdk::erc20::ERC20Action;
+use hyllar::HyllarAction;
 use sdk::utils::parse_contract_input;
 use sdk::{
     Blob, BlobIndex, ContractAction, ContractInput, Digestable, HyleContract, RunResult,
@@ -66,7 +66,7 @@ impl HyleContract for Amm {
                 // Check that a blob for the transfer exists for first token in swap
                 execution_ctx.is_in_callee_blobs(
                     &ContractName(pair.0.clone()),
-                    ERC20Action::TransferFrom {
+                    HyllarAction::TransferFrom {
                         owner: execution_ctx.caller.0.clone(),
                         recipient: execution_ctx.contract_name.0.clone(),
                         amount: from_amount,
@@ -75,7 +75,7 @@ impl HyleContract for Amm {
                 // Check that a blob for the transfer exists for second token in swap
                 execution_ctx.is_in_callee_blobs(
                     &ContractName(pair.1.clone()),
-                    ERC20Action::Transfer {
+                    HyllarAction::Transfer {
                         recipient: execution_ctx.caller.0.clone(),
                         amount: to_amount,
                     },
@@ -86,7 +86,7 @@ impl HyleContract for Amm {
                 // Check that a blob for the transfer exists for first token in pair
                 execution_ctx.is_in_callee_blobs(
                     &ContractName(pair.0.clone()),
-                    ERC20Action::TransferFrom {
+                    HyllarAction::TransferFrom {
                         owner: execution_ctx.caller.0.clone(),
                         recipient: execution_ctx.contract_name.0.clone(),
                         amount: amounts.0,
@@ -95,7 +95,7 @@ impl HyleContract for Amm {
                 // Check that a blob for the transfer exists for second token in pair
                 execution_ctx.is_in_callee_blobs(
                     &ContractName(pair.1.clone()),
-                    ERC20Action::TransferFrom {
+                    HyllarAction::TransferFrom {
                         owner: execution_ctx.caller.0.clone(),
                         recipient: execution_ctx.contract_name.0.clone(),
                         amount: amounts.1,
