+++ crates/contracts/hydentity/Cargo.toml
@@ -44,6 +44,7 @@ client-sdk = { workspace = true, features = ["risc0", "rest", "indexer"] }
 clap = { version = "4.5", features = ["derive"] }
 tokio = { version = "1.42.0", features = ["full", "tracing"] }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+mockall = "0.13.0"
 
 [features]
 default = []
