+++ tests/amm_test.rs
@@ -24,10 +24,11 @@ mod e2e_amm {
         client::{register_identity, verify_identity},
         Hydentity,
     };
-    use hyle_contract_sdk::{erc20::ERC20, guest, ContractInput, ContractName, HyleOutput};
+    use hyle_contract_sdk::{guest, ContractInput, ContractName, HyleOutput};
     use hyle_contracts::{AMM_ELF, HYDENTITY_ELF, HYLLAR_ELF};
     use hyllar::{
         client::{approve, transfer},
+        erc20::ERC20,
         Hyllar, FAUCET_ID,
     };
