+++ tests/hyllar_test.rs
@@ -16,9 +16,9 @@ mod e2e_hyllar {
         client::{register_identity, verify_identity},
         Hydentity,
     };
-    use hyle_contract_sdk::{erc20::ERC20, guest, ContractInput, ContractName, HyleOutput};
+    use hyle_contract_sdk::{guest, ContractInput, ContractName, HyleOutput};
     use hyle_contracts::{HYDENTITY_ELF, HYLLAR_ELF};
-    use hyllar::{client::transfer, Hyllar, FAUCET_ID};
+    use hyllar::{client::transfer, erc20::ERC20, Hyllar, FAUCET_ID};
 
     use super::*;
